crawler/requirements.txt
germansentiment
requests
bs4
pymongo
python-dotenv

fastapi
uvicorn
